tools/requirements.txt
publicsuffixlist==0.*
PyYAML==5.*
requests==2.*
six==1.*
tqdm==4.*
lxml==4.*
